Running through simulation
It is likely penalizing regular images, so it is likely not learning but also we do not have any experiments set up yet to even test it learning with hazards. Need to place hazards. Also see if we can use another reward function if we cant fix the anomaly detection algorithm, also train the anomaly detection algorithm more to really see.